Make IQ work when keyboard is already shown
Make IQKeyboard work when the keyboard is already shown and the
UITextField/UITextView becomes first responder (in this case
UIKeyboardWillShowNotification is not fired). Also IQ before this
change didn't work for me for UITextView (worked only for UITextField).
After the change IQ is working well both for UITextField and UITextView.